+++ GeeksDirectory.Services/Commands/RegisterProfileCommand.cs
@@ -6,7 +6,7 @@ using MediatR;
 
 namespace GeeksDirectory.Services.Commands
 {
-    public class RegisterProfileCommand : IRequest<Result<int>>
+    public class RegisterProfileCommand : IRequest<Result<long>>
     {
         public readonly CreateGeekProfileModel Profile;
