+++ GeeksDirectory.Services/Commands/RegisterSkillCommand.cs
@@ -6,13 +6,13 @@ using MediatR;
 
 namespace GeeksDirectory.Services.Commands
 {
-    public class RegisterSkillCommand : IRequest<Result<int>>
+    public class RegisterSkillCommand : IRequest<Result<long>>
     {
-        public readonly int ProfileId;
-        public readonly int SkillId;
+        public readonly long ProfileId;
+        public readonly long SkillId;
         public readonly SkillModel Skill;
 
-        public RegisterSkillCommand(int profileId, SkillModel skill) => 
+        public RegisterSkillCommand(long profileId, SkillModel skill) => 
             (this.ProfileId, this.Skill) = (profileId, skill);
     }
 }
