+++ GeeksDirectory.Services/Commands/UpdateProfileCommand.cs
@@ -6,12 +6,12 @@ using MediatR;
 
 namespace GeeksDirectory.Services.Commands
 {
-    public class UpdateProfileCommand : IRequest<Result<int>>
+    public class UpdateProfileCommand : IRequest<Result<long>>
     {
         public readonly GeekProfileModel Profile;
-        public readonly int ProfileId;
+        public readonly long ProfileId;
 
-        public UpdateProfileCommand(GeekProfileModel profile, int profileId) =>
+        public UpdateProfileCommand(GeekProfileModel profile, long profileId) =>
             (this.Profile, this.ProfileId) = (profile, profileId);
     }
 }
